Renamed Vector::dot to Vector::cos, and implement normal Vector::dot
For unit vectors (which were all I was originally going to support)
the dot product is the same as the cosine.  When I added support for
scaling the spaces, I added normalization for dot product, but should
have instead added 'cos' as a distinct function.

This is corrected, and breaks backward compatibility for scaled
spaces, but there aren't many users yet and it's more important to
use official terminology.

refs #2