Add support for projecting PDL vectors
The m3s_read_vector_from_sv function now returns a type code to
indicate what type it discovered, which simplifies the code that
needs to produce matching type of value during project functions.

For the moment, I'm calling out to perl-side for the PDL operations
because the normal XS way of calling PDL requires a versioned
struct from a header file, which would make compilation dependent
on having a current PDL header, and introduce a dependency on the
order the modules were installed.