Fix handling of vectors from hashrefs, and return value of *_inplace
One of the vector formats is { x => $x, y => $y, z => $z }, but
the project_* family of functions wasn't respecting that return
type.  Now it correctly returns a hashref for project, and modifies
the scalars of the existing hash for project_inplace.

I also fixed a bug where I was returning the PDL array from the
project_inplace family of methods, instead of returning $self.