Re-work Projection object / API
I decided to use the shorter names 'frustum' and 'perspective' as exported
functions, and split up the methods that get a GL matrix so that you can
choose a vector of floats or vector of doubles.  I gave up the optimization
of passing an existing scalar buffer into the function since it complicates
the API and is unlikely to result in any measurable speedup.