More renames in projection API, and tests for packing functions
I switched to have something like a 'matrix' attribute, but instead
of returning an object, return alternate representations indicated
by a suffix, such as _colmajor, _pack_float, or _pack_double.

Also added unit tests for the packing behavior.  Surprisingly the
bytes from packing with C are not the same as the bytes when packing
with 'pack("f*",...)', so I had to test the result of unpacking it.